refactor : moved articleRead component into Aricle>Read folder - reason : for gathering CRUD into same folder
